refactor: Dto 역할에 맞게 Column의 패키지 수정
